refactor(chat): thin chat_service.dart further — extract 3 more concerns to parts
Continue decomposing the god file with the part-file pattern (zero behaviour
change). Three more cohesive concerns moved into `part of` extensions:

- chat_service_group_settings.dart — per-group-character RAG/memory tuning +
  per-character Author's Note + per-character system-prompt overrides.
- chat_service_evolution.dart — character-evolution accessors and the manual
  edit/reset/group-load helpers (evolved personality/scenario deltas).
- chat_service_sillytavern.dart — SillyTavern chat JSON import/export.

The backing fields (e.g. `_groupEvolutionCounts`, `_evolvedPersonalities`) stay
in the main library — extensions cannot declare instance fields — so only methods
and getters moved.

Important caveat learned and handled: extension members are statically dispatched,
so a method moved to an extension can no longer be overridden by
`FakeChatService implements ChatService`. `characterEvolutionCount` is overridden
by that test fake, so it is deliberately kept as an in-class instance getter (with
a comment explaining why); the full unit suite (1334) and the golden suite both
pass, confirming no fake/golden path hits a moved member.

chat_service.dart: 10,506 → 9,993 lines across the four extractions so far.
flutter analyze clean (0 warnings).
